Moved sound store into store folder
