+++ frontend/src/Views/LevelView/QuestionView.vue
@@ -6,7 +6,7 @@ import type Teammate from '@/types/Teammate'
 import { storeToRefs } from 'pinia'
 import { computed, ref, toRefs, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { useSoundStore } from '@/composables/useSound'
+import { useSoundStore } from '@/stores/useSoundStore'
 import { useGame } from '@/stores/gameStore'
 import AnswerCard from './AnswerCard.vue'
 import BonusCard from './BonusCard.vue'
