+++ frontend/src/components/SharedComponents/LanguageButton.vue
@@ -39,7 +39,7 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { useSoundStore } from '@/composables/useSound'
+import { useSoundStore } from '@/stores/useSoundStore'
 
 const { playButtonSound } = useSoundStore()
 const { locale } = useI18n()
